engine: move a bunch of miscellaneous types to good homes
