+++ engine/esm/transpiled.js
@@ -102,149 +102,6 @@ import { TourEditTab } from "./tours/tour_edit.js";
 import { Folder } from "./folder.js";
 
 
-// wwtlib.PointType
-
-export var PointType = {
-  move: 0,
-  line: 1,
-  dash: 2,
-  start: 3
-};
-
-registerType("PointType", PointType);
-registerEnum("PointType", PointType);
-
-// wwtlib.Classification
-//
-// This was defined in `IPlace.cs`, which we've folded into `interfaces.js`.
-
-export var Classification = {
-  star: 1,
-  supernova: 2,
-  blackHole: 4,
-  neutronStar: 8,
-  doubleStar: 16,
-  multipleStars: 32,
-  asterism: 64,
-  constellation: 128,
-  openCluster: 256,
-  globularCluster: 512,
-  nebulousCluster: 1024,
-  nebula: 2048,
-  emissionNebula: 4096,
-  planetaryNebula: 8192,
-  reflectionNebula: 16384,
-  darkNebula: 32768,
-  giantMolecularCloud: 65536,
-  supernovaRemnant: 131072,
-  interstellarDust: 262144,
-  quasar: 524288,
-  galaxy: 1048576,
-  spiralGalaxy: 2097152,
-  irregularGalaxy: 4194304,
-  ellipticalGalaxy: 8388608,
-  knot: 16777216,
-  plateDefect: 33554432,
-  clusterOfGalaxies: 67108864,
-  otherNGC: 134217728,
-  unidentified: 268435456,
-  solarSystem: 536870912,
-  unfiltered: 1073741823,
-  stellar: 63,
-  stellarGroupings: 2032,
-  nebulae: 523264,
-  galactic: 133693440,
-  other: 436207616
-};
-
-registerType("Classification", Classification);
-registerEnum("Classification", Classification);
-
-// wwtlib.StockSkyOverlayTypes
-//
-// This was defined in `Tours/ISettings.cs`, which we've folded into `interfaces.js`.
-
-export var StockSkyOverlayTypes = {
-  empty: 0,
-  equatorialGrid: 1,
-  equatorialGridText: 2,
-  galacticGrid: 3,
-  galacticGridText: 4,
-  eclipticGrid: 5,
-  eclipticGridText: 6,
-  eclipticOverview: 7,
-  eclipticOverviewText: 8,
-  precessionChart: 9,
-  altAzGrid: 10,
-  altAzGridText: 11,
-  constellationFigures: 12,
-  constellationBoundaries: 13,
-  constellationFocusedOnly: 14,
-  constellationNames: 15,
-  constellationPictures: 16,
-  fadeToBlack: 17,
-  fadeToLogo: 18,
-  fadeToGradient: 19,
-  screenBroadcast: 20,
-  fadeRemoteOnly: 21,
-  skyGrids: 22,
-  constellations: 23,
-  solarSystemStars: 24,
-  solarSystemMilkyWay: 25,
-  solarSystemCosmos: 26,
-  solarSystemOrbits: 27,
-  solarSystemPlanets: 28,
-  solarSystemAsteroids: 29,
-  solarSystemLighting: 30,
-  solarSystemMinorOrbits: 31,
-  showEarthCloudLayer: 32,
-  showElevationModel: 33,
-  showAtmosphere: 34,
-  multiResSolarSystemBodies: 35,
-  auroraBorialis: 36,
-  earthCutAway: 37,
-  showSolarSystem: 38,
-  clouds8k: 39,
-  filedOfView: 40,
-  showISSModel: 41,
-  solarSystemCMB: 42,
-  mpcZone1: 43,
-  mpcZone2: 44,
-  mpcZone3: 45,
-  mpcZone4: 46,
-  mpcZone5: 47,
-  mpcZone6: 48,
-  mpcZone7: 49,
-  orbitFilters: 50
-};
-
-registerType("StockSkyOverlayTypes", StockSkyOverlayTypes);
-registerEnum("StockSkyOverlayTypes", StockSkyOverlayTypes);
-
-// wwtlib.DialogResult
-
-export var DialogResult = {
-  OK: 1
-};
-
-registerType("DialogResult", DialogResult);
-registerEnum("DialogResult", DialogResult);
-
-// wwtlib.Formatting
-
-export var Formatting = {
-  indented: 1
-};
-
-registerType("Formatting", Formatting);
-registerEnum("Formatting", Formatting);
-
-// GFX
-
-export function GFX() { }
-
-registerType("GFX", [GFX, null, null]);
-
 // The `Layer.fromXml` function, which needs to be aware of a bunch of layer
 // subclasses.
